Rename module `translate.py` to `api.py`
